Refactored upload authentication and granting
The authentication now has a preliminary API call with a web guard that invoke the Gate, via FormRequest. If authentication and validation passes the upload is stored in the queue and signaled as started. The JS now switch to tusd-client for doing the upload.
Moved also to rollupjs for packaging the library as gives a lower footprint